MyProjFolder/requirements.txt
azure-functions==1.13.2
pytest==8.3.3
jsonschema==4.23.0
